move functions that create relative coordinates to geometry source file, rename interaction_list.h to build_list.h, change hash_lut to global
